requirements.txt
joblib==0.14.1
numpy==1.18.2
Pillow==7.1.1
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
pytorch==1.4.0
